fixes so mSEM scenes work again, workaround for 3 channels in non-mosaic file, access functions for nscenes and scene bounding boxes, major changes to mSEM scene again to support ATUM ribbons and loading cz annotation files

+++ _pylibczi/pb_helpers.h
@@ -33,7 +33,7 @@ namespace pb_helpers {
       T* position = ptr;
       for (const auto& image : images_) {
           auto typedImage = pylibczi::ImageFactory::getDerived<T>(image);
-          int length = typedImage->length();
+          size_t length = typedImage->length();
           std::copy(typedImage->getRawPtr(), typedImage->getRawPtr()+length, position);
           position += length;
       }
